Prepped "Picture Worth 1,000 Words" First Image Drop.
The "Full-Annual-Panel" Chicago Police Department Calendar with transition Ᏼlend is better at explaining and effectuating the 13 Month Overlay Transition. i'm still marvelling at the nonobvious design needs, like the "hot focus" reqs on the 13s.. i'm very impressed. Surprised it couldn't be done with a "cooler" color, but that could be because it's public-facing, an introduction series, or could have only been more muted a few months in to the transition of thought, and isn't a big deal for the final runs. i think i would have done something lighter, less saturated, but the need is obvious, and it definitely has to keep drawing the eye toward a new "center" and focus point. Truly very impressed, if that is in fact what i think it was.