Hail Mary! Pass (almost ready)
Have to check thru the Work to discover the Easter Objections, and the issues with the Pax Calendar that the Catholic Church might have expressed. i don't understand y⁵et what anchors the scheduling of Easter, and how that affects this, and i have to before i can really release this with total confidence that eventually, even if Years late and on a very unmemorable Year, it will get adopted, on CORE and Ᏼranch Thermodynamic Grounds, with better catalysts than ever these days.